#24 [refactor] code refactoring

+++ src/main/java/org/crops/fitserver/domain/user/dto/request/UpdateUserRequest.java
@@ -33,7 +33,7 @@ public class UpdateUserRequest {
 
   private Integer projectCount;
 
-  private Integer activityHour;
+  private Short activityHour;
 
   private String introduce;
